Disable arrow keys in normal mode
Because my Ergodox keyboard has mapped the actual arrow keys to HJKL, it
will be easy for me to accidentally rely on the arrow keys instead of
using HJKL properly.

Not a big deal for my laptop because the arrow keys are far away on the
bottom right part of the keyboard.