Markdown improvements with autolist indentation
The autolist plugin makes it really easy to edit markdown files,
automatically inserting the - or 1 at the beginning of a new unordered
or ordered list, respectively.

To test it out, I updated the local TODO.md file with it, and it works
great! Just have to remember to make autolist depend on autopairs,
because autopairs clobbers some of the config.

Also, treesitter's default indentation for markdown files is really bad,
so we have to turn it off to get the proper indentation (where the first
character of the second line of an unordered list item is directly
underneath the first character of the first line).