Replace Mason elixirls with elixir-tools
These two don't play well together :( If they play well in the future,
I'll migrate back to Mason.

elixir-tools installs its LSP binaries in $HOME/.cache, so at least it's
smart enough not to install the same exact LSP over and over again.
It'll only install a different version if it sees that my current
project is using a different version of OTP, which is fair.

If I think of a good reason to enable nextls or projectionist, the
config will make it easy to do so, but I can't think of a good reason at
this time.