Disable function signature help menu
When I'm writing a function, it's distracting and disorienting to have
the function's signature pop above my cursor with a help menu. It makes
it harder to read the line above my cursor, which is often the line I
just finished writing. If the help menu can be configured to go *below*
my cursor, or if I get into a language where long function signatures
are common, then this feature could be useful. But for now, not useful.