notify before using LSP as fallback formatter
I want to know about the configured formatter failing, because the LSP's
formatting behavior might be different, and I might not notice the
difference.

I've also added a comment to illustrate what other options are possible,
such as turning off the fallback entirely with `false`. I'll leave the
fallback up for now because at least I'm getting notifications!

This was motivated by initially being confused about whether `mix` was
trimming trailing whitespace or not. Turns out it *was* trimming
trailing whitespace, just not for comments.